requirements.txt
discord.py==1.7.1
NudeNet==2.0.9
Pillow==8.2.0
pynacl==1.4.0
PyYAML==5.4.1
requests==2.25.1
python-dotenv==0.18.0